合并元素反应
